requirements.txt
discord.py==1.5.1
asyncio==3.4.3
google-api-python-client==1.12.5
googletrans==3.0.0
jikanpy==4.2.2
requests==2.24.0
imgurpython==1.1.7
qrcode==6.1
pokedex.py==1.1.2
pypokedex==1.5.1
pokebase==1.3.0